Extract component for ‘Residence’ section
