+++ src/app/features/types.tsx
@@ -37,7 +37,7 @@ export type BrkDataError = {
   error: string
 }
 
-type ImportAdres = {
+export type ImportAdres = {
   hsltr: string | null
   hsnr: string
   postcode: string
